requirements.txt
Flask==3.0.0
pdfplumber
python-dateutil
psycopg2-binary